+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This release contains new migrations and requires to run the `migrate` command a
 ## Added
 
 + [#216](https://github.com/luyadev/luya-module-admin/issues/216) File manager file detail view provides option to tag files.
++ [#259](https://github.com/luyadev/luya-module-admin/pull/259) SelectRelationActiveQuery supports related i18n label fields
 + [#253](https://github.com/luyadev/luya-module-admin/pull/253) Added command action to reset password for users via cli.
 
 ## 1.2.3 (21. November 2018)